requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm==0.9.12
albumentations==1.3.1
einops==0.7.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
Pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0
scipy>=1.10.0
tensorboard>=2.7.0
scikit-learn>=0.24.2
